firmware updates

+++ firmware_17xx/synth/wtosc.h
@@ -24,6 +24,6 @@ struct wtosc_s
 void wtosc_init(struct wtosc_s * o, int8_t channel, uint16_t controlData);
 void wtosc_setSampleData(struct wtosc_s * o, int16_t * data, uint16_t sampleCount);
 void wtosc_setParameters(struct wtosc_s * o, uint16_t cv, uint16_t aliasing);
-uint32_t wtosc_update(struct wtosc_s * o);
+uint16_t wtosc_update(struct wtosc_s * o);
 
 #endif
